Refactor Label.
