+++ client/src/components/Label/index.js
@@ -1,0 +1,3 @@
+import Label from './Label.jsx';
+
+export default Label;
